Resolve conversation (#20)
* uninstall new version react-router/install react-router-dom

* change Participants to string

* use <Link> from React instead of <a>

* change to correct button name

* resolve duplicate files

* make feed responsive
