+++ package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
-    "react-router": "^7.9.4",
+    "react-router-dom": "^6.30.1",
     "styled-components": "^6.1.19"
   },
   "devDependencies": {
