+++ src/Components/feed-components/Card.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { useState } from "react";
+import { useNavigate } from "react-router-dom";
 import {
   Container,
   Hero,
@@ -16,12 +17,13 @@ import {
   LocationRow,
   LocationInfo,
   JoinButton,
+  ElypsisText,
 } from "../styled/feed-style-comp/Card.styled.jsx";
-import ParticipantsIcon from "../../assets/Icons/participants.svg?react";
-import LocationIcon from "../../assets/Icons/map-pin-green.svg?react";
-import { ElypsisText } from "../styled/feed-style-comp/Card.styled.jsx";
+import ParticipantsIcon from "../../assets/icons/participants.svg?react";
+import LocationIcon from "../../assets/icons/map-pin-green.svg?react";
 
 export default function Card({
+  id,
   image,
   date,
   priceLabel,
@@ -30,12 +32,26 @@ export default function Card({
   host,
   participants,
   location,
-  onJoin,
+  onJoin = () => {},
 }) {
   // new: manage joined state locally
   const [joined, setJoined] = useState(false);
+  const navigate = useNavigate();
 
-  const handleToggleJoin = () => {
+  const goToDetail = () => {
+    if (!id) return;
+    navigate(`/activity/${id}`);
+  };
+
+  const onKeyGoToDetail = (e) => {
+    if (e.key === "Enter" || e.key === " ") {
+      e.preventDefault();
+      goToDetail();
+    }
+  };
+
+  const handleToggleJoin = (e) => {
+    e.stopPropagation();
     const next = !joined;
     setJoined(next);
     // call external handler when joined; you can modify to call onJoin/onCancel separately
@@ -43,7 +59,14 @@ export default function Card({
   };
 
   return (
-    <Container>
+    <Container
+      onClick={goToDetail}
+      onKeyDown={onKeyGoToDetail}
+      role="button"
+      tabIndex={0}
+      style={{ cursor: "pointer" }}
+      aria-label={`Open details for ${title}`}
+    >
       <Hero>
         {image && <HeroImage src={image} alt={title} />}
         <CornerChips>{date}</CornerChips>
@@ -66,17 +89,7 @@ export default function Card({
           </HostInfo>
 
           <Participants title="Participants">
-            {typeof ParticipantsIcon === "string" ? (
-              <img
-                src={ParticipantsIcon}
-                width={20}
-                height={20}
-                alt=""
-                aria-hidden
-              />
-            ) : (
-              <ParticipantsIcon width={20} height={20} aria-hidden />
-            )}
+            <ParticipantsIcon width={20} height={20} aria-hidden />
             <span style={{ marginLeft: 6 }}>
               {participants.count}/{participants.max}
             </span>
@@ -98,6 +111,7 @@ export default function Card({
             )}
             <ElypsisText style={{ marginLeft: 6 }}>{location}</ElypsisText>
           </LocationInfo>
+
           <JoinButton
             joined={joined ? 1 : 0}
             onClick={handleToggleJoin}
