+++ src/Components/feed-components/Footer.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Link } from "react-router";
+import { Link } from "react-router-dom";
 import BondiLogo from "../../assets/Icons/second-logo.svg?react";
 import {
   FooterInner,
