+++ src/main.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./main.css";
 
-import { BrowserRouter, Routes } from "react-router";
+import { BrowserRouter, Routes } from "react-router-dom";
 import Router from "./Router.jsx";
 
 const root = document.getElementById("root");
