+++ src/pages/Feed.jsx
@@ -1,8 +1,8 @@
 import React from "react";
-import NavBar from "../Components/feed-components/NavBar.jsx";
-import Footer from "../Components/feed-components/Footer.jsx";
-import Filter from "../components/Filter.jsx";
-import Card from "../components/Card.jsx";
+import NavBar from "../components/feed-components/NavBar.jsx";
+import Footer from "../components/feed-components/Footer.jsx";
+import Filter from "../components/feed-components/Filter.jsx";
+import Card from "../components/feed-components/Card.jsx";
 import YogaImage from "../assets/images/yoga.jpg";
 import HostYoga from "../assets/images/olly-yoga.jpg";
 import HostVolleyBall from "../assets/images/avatar.png";
@@ -15,11 +15,15 @@ import TriviaImage from "../assets/images/trivia.jpg";
 import HostTrivia from "../assets/images/clara.jpg";
 import BoardGameImage from "../assets/images/boardgame.jpg";
 import HostBoardGame from "../assets/images/elias.jpg";
-import { MainContainer } from "../components/styled/Middle.styled.jsx";
-import { LayoutGrid } from "../components/styled/Grid.styled.jsx";
-import { GridContainer } from "../components/styled/Grid.styled.jsx";
-import { SectionHeader } from "../components/styled/Feed.styled.jsx";
-import { LoadMoreButton } from "../components/styled/Feed.styled.jsx";
+import { MainContainer } from "../components/styled/activity-detail-comp/Middle.styled.jsx";
+import {
+  LayoutGrid,
+  GridContainer,
+} from "../components/styled/feed-style-comp/Grid.styled.jsx";
+import {
+  SectionHeader,
+  LoadMoreButton,
+} from "../components/styled/feed-style-comp/Feed.styled.jsx";
 
 export default function Feed() {
   const handleApply = (filters) => {
@@ -128,15 +132,11 @@ export default function Feed() {
   return (
     <>
       <NavBar />
-      <MainContainer
-        style={{ display: "flex", gap: 40, alignItems: "flex-start" }}
-      >
+      <MainContainer>
         <LayoutGrid>
-          <div style={{ width: 320 }}>
-            <Filter onApply={handleApply} />
-          </div>
+          <Filter onApply={handleApply} />
 
-          <div style={{ flex: 1 }}>
+          <div>
             <SectionHeader>
               <h1>Discover Activities</h1>
               <p>Find amazing activities happening around you</p>
