+++ src/pages/RedirectToLogin.jsx
@@ -1,5 +1,5 @@
 import { useEffect } from "react";
-import { useNavigate } from "react-router";
+import { useNavigate } from "react-router-dom";
 
 function RedirectToLogin() {
   let navigate = useNavigate();
